backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyPDF2>=3.0.0
google-generativeai==0.3.2
python-dotenv>=1.0.0